LayoutTests/canvas/philip/tests/2d.imageData.get.zero-expected.txt

< [index] >
2d.imageData.get.zero
getImageData() throws INDEX_SIZE_ERR if size is zero
References: 2d.imageData.getcreate.zero
Actual output:
Passed
